feat: batch inventory tracking with FEFO consumption
- Add ingredient_batches table (quantity, expiry_date, FK to ingredients)
- New API endpoints: POST /stock-in, GET/PUT /batches, POST /batches/delete
- Stock In modal: add multiple ingredient batches with optional expiry dates in one step
- Edit modal: view and manage individual batches per ingredient
- Inventory table: sortable columns (name, expiry), expiry badge with color coding
- Cook meal now consumes batches FEFO (earliest expiry first, nulls last)
- Stock recomputed as SUM of batches on any batch mutation (no drift)
- price field changed INTEGER across model, schema, and recipe cost display
- EarliestExpiry surfaced on inventory list for expiry-based sorting

+++ backend/handlers/recipe.go
@@ -109,11 +109,11 @@ func (h *RecipeHandler) GetRecipeIngredients(w http.ResponseWriter, r *http.Requ
 	defer rows.Close()
 
 	type IngredientDetail struct {
-		IngredientID int     `json:"ingredient_id"`
-		Name         string  `json:"name"`
-		Quantity     int     `json:"quantity"`
-		Price        float64 `json:"price"`
-		IsTracked    bool    `json:"is_tracked"`
+		IngredientID int    `json:"ingredient_id"`
+		Name         string `json:"name"`
+		Quantity     int    `json:"quantity"`
+		Price        int    `json:"price"`
+		IsTracked    bool   `json:"is_tracked"`
 	}
 
 	var ingredients []IngredientDetail
